# StringDictionarySlim instead of Dictionary/StructDictionary