… Jint/Runtime/Environments/DeclarativeEnvironmentRecord.cs
… Jint/Runtime/StructDictionary.cs