I have replaced the StructDictionary with slightly modified [DictionarySlim](https://github.com/dotnet/corefxlab/pull/2458). This version uses string keys directly (hence the name StringDictionarySlim) and doesn't check for collisions as we don't need to detect multi-threaded access.

I also removed separate dictionary _intrinsicProperties as so far I've seen need, we can probably add back later if needed.

Performance results generally seem to be within the waving my laptop does, but the greatest benefit can be seen in UncacheableExpressionsBenchmark where more realistic object mapping happens against the new dictionary type.

## Esprima.Benchmark.DromaeoBenchmark

| **Diff**|Method|FileName|Mean|Gen 0/1k Op|Gen 1/1k Op|Gen 2/1k Op|Allocated Memory/Op|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------|
| Old |Run|dromaeo-3d-cube|72.63 ms|1250.0000|250.0000|-|7366 KB|
| **New** |	|	| **72.00 ms (-1%)** | **1285.7143 (+3%)** | **285.7143 (+14%)** | **-** | **7269.35 KB (-1%)** |
| Old |Run|dromaeo-core-eval|14.60 ms|62.5000|-|-|297.95 KB|
| **New** |	|	| **14.77 ms (+1%)** | **62.5000 (0%)** | **-** | **-** | **297.64 KB (0%)** |
| Old |Run|dromaeo-object-array|162.35 ms|42000.0000|1750.0000|750.0000|178311.04 KB|
| **New** |	|	| **162.49 ms (0%)** | **42000.0000 (0%)** | **1750.0000 (0%)** | **750.0000 (0%)** | **178094.41 KB (0%)** |
| Old |Run|droma(...)egexp [21]|869.95 ms|51000.0000|29000.0000|19000.0000|412014.58 KB|
| **New** |	|	| **878.73 ms (+1%)** | **50000.0000 (-2%)** | **29000.0000 (0%)** | **19000.0000 (0%)** | **408868.92 KB (-1%)** |
| Old |Run|droma(...)tring [21]|1,174.86 ms|219000.0000|176000.0000|173000.0000|1392945.34 KB|
| **New** |	|	| **1,165.84 ms (-1%)** | **217000.0000 (-1%)** | **175000.0000 (-1%)** | **173000.0000 (0%)** | **1394102.99 KB (0%)** |


## Jint.Benchmark.ArrayBenchmark

| **Diff**|Method|N|Mean|Gen 0/1k Op|Gen 1/1k Op|Gen 2/1k Op|Allocated Memory/Op|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------|
| Old |Slice|100|429.3 us|161.1328|-|-|660.16 KB|
| **New** |	|	| **446.7 us (+4%)** | **160.6445 (0%)** | **-** | **-** | **659.38 KB (0%)** |
| Old |Concat|100|461.8 us|175.7813|-|-|720.31 KB|
| **New** |	|	| **463.3 us (0%)** | **175.2930 (0%)** | **-** | **-** | **718.75 KB (0%)** |
| Old |Unshift|100|18,291.9 us|3562.5000|-|-|14672.66 KB|
| **New** |	|	| **18,186.7 us (-1%)** | **3562.5000 (0%)** | **-** | **-** | **14671.88 KB (0%)** |
| Old |Push|100|10,113.7 us|343.7500|-|-|1438.28 KB|
| **New** |	|	| **10,054.4 us (-1%)** | **343.7500 (0%)** | **-** | **-** | **1437.5 KB (0%)** |
| Old |Index|100|11,500.0 us|390.6250|-|-|1637.5 KB|
| **New** |	|	| **11,431.0 us (-1%)** | **390.6250 (0%)** | **-** | **-** | **1636.72 KB (0%)** |
| Old |Map|100|2,635.9 us|488.2813|-|-|2011.72 KB|
| **New** |	|	| **2,530.7 us (-4%)** | **488.2813 (0%)** | **-** | **-** | **2009.38 KB (0%)** |
| Old |Apply|100|603.4 us|188.4766|-|-|774.22 KB|
| **New** |	|	| **583.6 us (-3%)** | **188.4766 (0%)** | **-** | **-** | **773.44 KB (0%)** |
| Old |JsonStringifyParse|100|4,598.3 us|1273.4375|-|-|5242.19 KB|
| **New** |	|	| **4,511.3 us (-2%)** | **1273.4375 (0%)** | **-** | **-** | **5233.59 KB (0%)** |
| Old |FilterWithString|100|28,951.9 us|3187.5000|-|-|13182.81 KB|
| **New** |	|	| **27,827.2 us (-4%)** | **3187.5000 (0%)** | **-** | **-** | **13180.47 KB (0%)** |


## Jint.Benchmark.UncacheableExpressionsBenchmark

| **Diff**|Method|N|Mean|Gen 0/1k Op|Gen 1/1k Op|Gen 2/1k Op|Allocated Memory/Op|
|------- |-------|-------|-------:|-------:|-------:|-------:|-------|
| Old |Benchmark|500|287.9 ms|33500.0000|-|-|135.1 MB|
| **New** |	|	| **251.6 ms (-13%)** | **29500.0000 (-12%)** | **-** | **-** | **119.66 MB (-11%)** |